報告書功能完成及功能修正
1.新增了物料項目及排放源預設項目
2.排放源新增返回及完成按鈕指向位置錯誤修改
3.刪除了不必要的空格
4.修正了化糞池排放量計算
5.報告書功能完成

+++ Models/Material.cs
@@ -20,7 +20,7 @@ namespace Carbon_inventory_platform.Models
         public string EmissionPattern { get; set; } = "";
 
         [Display(Name = "CO2排放係數")]
-        public float CO2CEF { get; set; } = 0;
+        public double CO2CEF { get; set; } = 0;
 
         [Display(Name = "CO2不確定性95%信賴區間下限")]
         public float CO2ULL { get; set; } = 0;
@@ -29,7 +29,7 @@ namespace Carbon_inventory_platform.Models
         public float CO2UUL { get; set; } = 0;
 
         [Display(Name = "CH4排放係數")]
-        public float CH4CEF { get; set; } = 0;
+        public double CH4CEF { get; set; } = 0;
 
         [Display(Name = "CH4不確定性95%信賴區間下限")]
         public float CH4ULL { get; set; } = 0;
@@ -38,7 +38,7 @@ namespace Carbon_inventory_platform.Models
         public float CH4UUL { get; set; } = 0;
 
         [Display(Name = "N2O排放係數")]
-        public float N2OCEF { get; set; } = 0;
+        public double N2OCEF { get; set; } = 0;
 
         [Display(Name = "N2O不確定性95%信賴區間下限")]
         public float N2OULL { get; set; }=0;
